Refactor AccountsWorkspace, GoalsWorkspace, and TransactionsWorkspace components: Simplify state management, remove unused types and functions, and enhance error handling. Integrate new form components for better structure and maintainability.
