+++ components/accounts-workspace.tsx
@@ -2,19 +2,10 @@
 
 import { startTransition, useEffect, useState } from "react";
 
-import { createBootstrapState } from "@/lib/app-state/bootstrap";
 import { defaultAccountTypes } from "@/lib/app-state/defaults";
 import { getAccountTotals } from "@/lib/domain/accounts";
 import { createIndexedDbRepositories } from "@/lib/repositories/indexeddb";
-import type {
-  Account,
-  AccountType,
-  IncomeType,
-  RiskComfort,
-  SalaryCycle,
-  UserProfile,
-} from "@/lib/types";
-import { Button } from "@/components/ui/button";
+import type { Account, AccountType } from "@/lib/types";
 import {
   Card,
   CardContent,
@@ -22,40 +13,15 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+
 import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
-import { Textarea } from "@/components/ui/textarea";
+  AccountForm,
+  type AccountFormState,
+  defaultAccountForm,
+} from "./accounts/account-form";
+import { AccountList } from "./accounts/account-list";
 
 const repositories = createIndexedDbRepositories();
-
-type AccountFormState = {
-  id?: string;
-  name: string;
-  type: AccountType;
-  institutionName: string;
-  openingBalance: string;
-  notes: string;
-};
-
-const defaultAccountForm: AccountFormState = {
-  name: "",
-  type: "cash",
-  institutionName: "",
-  openingBalance: "0",
-  notes: "",
-};
-
-function buildTimestamp() {
-  return new Date().toISOString();
-}
 
 function formatCurrency(amount: number) {
   return new Intl.NumberFormat("en-UG", {
@@ -72,17 +38,10 @@ function toInstitutionType(type: AccountType): Account["institutionType"] {
   return "other";
 }
 
-const accountTypeLabels: Record<AccountType, string> = {
-  cash: "Cash",
-  mobile_money: "Mobile Money",
-  bank: "Bank Account",
-  sacco: "SACCO",
-  investment: "Investment",
-  debt: "Debt / Obligation",
-};
-
 export function AccountsWorkspace() {
-  const [profile, setProfile] = useState<UserProfile | null>(null);
+  const [profile, setProfile] = useState<
+    { id: string } | null
+  >(null);
   const [accounts, setAccounts] = useState<Account[]>([]);
   const [accountForm, setAccountForm] = useState<AccountFormState>(defaultAccountForm);
   const [editingAccountId, setEditingAccountId] = useState<string | null>(null);
@@ -105,9 +64,7 @@ export function AccountsWorkspace() {
         setAccounts([]);
       }
     } catch (loadError) {
-      setError(
-        loadError instanceof Error ? loadError.message : "Unable to load accounts.",
-      );
+      setError(loadError instanceof Error ? loadError.message : "Unable to load accounts.");
     } finally {
       setIsLoading(false);
     }
@@ -121,18 +78,17 @@ export function AccountsWorkspace() {
 
   async function handleAccountSubmit(event: React.FormEvent<HTMLFormElement>) {
     event.preventDefault();
-
     if (!profile) return;
 
     setIsSubmitting(true);
     setError(null);
 
     try {
-      const timestamp = buildTimestamp();
+      const timestamp = new Date().toISOString();
       const accountId = editingAccountId ?? `account:${crypto.randomUUID()}`;
       const openingBalance = Number(accountForm.openingBalance);
 
-      const nextAccount: Account = {
+      await repositories.accounts.upsert({
         id: accountId,
         userId: profile.id,
         name: accountForm.name.trim(),
@@ -145,16 +101,13 @@ export function AccountsWorkspace() {
         isArchived: false,
         createdAt: accounts.find((a) => a.id === accountId)?.createdAt ?? timestamp,
         updatedAt: timestamp,
-      };
+      });
 
-      await repositories.accounts.upsert(nextAccount);
       setAccountForm(defaultAccountForm);
       setEditingAccountId(null);
       await loadWorkspace();
     } catch (submitError) {
-      setError(
-        submitError instanceof Error ? submitError.message : "Unable to save account.",
-      );
+      setError(submitError instanceof Error ? submitError.message : "Unable to save account.");
     } finally {
       setIsSubmitting(false);
     }
@@ -163,7 +116,6 @@ export function AccountsWorkspace() {
   function beginAccountEdit(account: Account) {
     setEditingAccountId(account.id);
     setAccountForm({
-      id: account.id,
       name: account.name,
       type: account.type,
       institutionName: account.institutionName ?? "",
@@ -241,172 +193,20 @@ export function AccountsWorkspace() {
           </div>
 
           <div className="grid gap-5 lg:grid-cols-[0.9fr_1.1fr]">
-            <Card className="border-border/40 shadow-none">
-              <CardHeader>
-                <CardTitle className="text-base">
-                  {editingAccountId ? "Edit account" : "Add account"}
-                </CardTitle>
-                <CardDescription>
-                  {editingAccountId
-                    ? "Update the details for this account."
-                    : "Add a new account to track."}
-                </CardDescription>
-              </CardHeader>
-              <CardContent>
-                <form className="grid gap-4" onSubmit={handleAccountSubmit}>
-                  <div className="grid gap-2">
-                    <Label htmlFor="account-name">Account name</Label>
-                    <Input
-                      id="account-name"
-                      value={accountForm.name}
-                      onChange={(e) =>
-                        setAccountForm((c) => ({ ...c, name: e.target.value }))
-                      }
-                      placeholder="e.g. MTN Mobile Money"
-                      required
-                    />
-                  </div>
+            <AccountForm
+              accountTypes={defaultAccountTypes}
+              form={accountForm}
+              editingId={editingAccountId}
+              isSubmitting={isSubmitting}
+              onFormChange={setAccountForm}
+              onSubmit={(e) => void handleAccountSubmit(e)}
+              onCancelEdit={() => {
+                setEditingAccountId(null);
+                setAccountForm(defaultAccountForm);
+              }}
+            />
 
-                  <div className="grid gap-2">
-                    <Label htmlFor="account-type">Account type</Label>
-                    <Select
-                      value={accountForm.type}
-                      onValueChange={(value) =>
-                        setAccountForm((c) => ({ ...c, type: value as AccountType }))
-                      }
-                    >
-                      <SelectTrigger id="account-type">
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        {defaultAccountTypes.map((type) => (
-                          <SelectItem key={type} value={type}>
-                            {accountTypeLabels[type]}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div className="grid gap-2">
-                    <Label htmlFor="institution-name">Institution name</Label>
-                    <Input
-                      id="institution-name"
-                      value={accountForm.institutionName}
-                      onChange={(e) =>
-                        setAccountForm((c) => ({ ...c, institutionName: e.target.value }))
-                      }
-                      placeholder="Optional — e.g. Stanbic Bank"
-                    />
-                  </div>
-
-                  <div className="grid gap-2">
-                    <Label htmlFor="opening-balance">Opening balance (UGX)</Label>
-                    <Input
-                      id="opening-balance"
-                      inputMode="decimal"
-                      value={accountForm.openingBalance}
-                      onChange={(e) =>
-                        setAccountForm((c) => ({ ...c, openingBalance: e.target.value }))
-                      }
-                      required
-                    />
-                  </div>
-
-                  <div className="grid gap-2">
-                    <Label htmlFor="account-notes">Notes</Label>
-                    <Textarea
-                      id="account-notes"
-                      value={accountForm.notes}
-                      onChange={(e) =>
-                        setAccountForm((c) => ({ ...c, notes: e.target.value }))
-                      }
-                      placeholder="Optional"
-                      className="min-h-20"
-                    />
-                  </div>
-
-                  <div className="flex flex-wrap gap-2">
-                    <Button disabled={isSubmitting} type="submit" size="sm">
-                      {isSubmitting
-                        ? "Saving..."
-                        : editingAccountId
-                          ? "Update account"
-                          : "Add account"}
-                    </Button>
-                    {editingAccountId ? (
-                      <Button
-                        type="button"
-                        variant="outline"
-                        size="sm"
-                        onClick={() => {
-                          setEditingAccountId(null);
-                          setAccountForm(defaultAccountForm);
-                        }}
-                      >
-                        Cancel
-                      </Button>
-                    ) : null}
-                  </div>
-                </form>
-              </CardContent>
-            </Card>
-
-            <Card className="border-border/40 shadow-none">
-              <CardHeader>
-                <CardTitle className="text-base">Your accounts</CardTitle>
-                <CardDescription>
-                  All accounts reconciled from opening balance and transaction history.
-                </CardDescription>
-              </CardHeader>
-              <CardContent className="grid gap-3">
-                {accounts.length === 0 ? (
-                  <div className="rounded-md border border-dashed border-border/50 px-4 py-8 text-sm text-muted-foreground">
-                    No accounts yet. Add your first account to get started.
-                  </div>
-                ) : (
-                  accounts
-                    .filter((a) => !a.isArchived)
-                    .map((account) => (
-                      <Card
-                        key={account.id}
-                        className="border-border/40 bg-muted/30 shadow-none"
-                      >
-                        <CardContent className="px-4 py-4">
-                          <div className="flex items-start justify-between gap-3">
-                            <div className="space-y-0.5">
-                              <div className="text-sm font-medium text-foreground">
-                                {account.name}
-                              </div>
-                              <div className="text-xs text-muted-foreground">
-                                {accountTypeLabels[account.type]}
-                                {account.institutionName
-                                  ? ` · ${account.institutionName}`
-                                  : ""}
-                              </div>
-                            </div>
-                            <Button
-                              size="sm"
-                              variant="ghost"
-                              className="h-7 text-xs"
-                              onClick={() => beginAccountEdit(account)}
-                            >
-                              Edit
-                            </Button>
-                          </div>
-                          <Separator className="my-3" />
-                          <div className="flex items-center justify-between text-sm">
-                            <span className="text-muted-foreground">Balance</span>
-                            <span className="font-medium tabular-nums">
-                              {formatCurrency(account.balance)}
-                            </span>
-                          </div>
-                        </CardContent>
-                      </Card>
-                    ))
-                )}
-              </CardContent>
-            </Card>
+            <AccountList accounts={accounts} onEdit={beginAccountEdit} />
           </div>
         </>
       ) : null}
