+++ components/goals-workspace.tsx
@@ -4,51 +4,19 @@ import { startTransition, useEffect, useMemo, useState } from "react";
 
 import { defaultGoalTypes } from "@/lib/app-state/defaults";
 import { reconcileAccountBalances } from "@/lib/domain/accounts";
-import { applyGoalTransactions, getGoalContributionPlan } from "@/lib/domain/goals";
+import { applyGoalTransactions } from "@/lib/domain/goals";
 import { getMonthSummary } from "@/lib/domain/summaries";
 import { createIndexedDbRepositories } from "@/lib/repositories/indexeddb";
-import type { Account, Goal, GoalType, Transaction, UserProfile } from "@/lib/types";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
+import type { Account, Goal, Transaction, UserProfile } from "@/lib/types";
+import { Card, CardContent } from "@/components/ui/card";
+
+import { GoalForm, type GoalFormState, defaultGoalForm } from "./goals/goal-form";
+import { GoalList } from "./goals/goal-list";
 
 const repositories = createIndexedDbRepositories();
 
-type GoalFormState = {
-  name: string;
-  goalType: GoalType;
-  targetAmount: string;
-  targetDate: string;
-  priority: string;
-  linkedAccountId: string;
-};
-
-const defaultGoalForm: GoalFormState = {
-  name: "",
-  goalType: "emergency_fund",
-  targetAmount: "",
-  targetDate: "",
-  priority: "1",
-  linkedAccountId: "",
-};
-
-function buildTimestamp() {
-  return new Date().toISOString();
+function sortGoals(goals: Goal[]) {
+  return [...goals].sort((a, b) => a.priority - b.priority);
 }
 
 function formatCurrency(amount: number) {
@@ -58,24 +26,6 @@ function formatCurrency(amount: number) {
     maximumFractionDigits: 0,
   }).format(amount);
 }
-
-function getCurrentMonth() {
-  return new Date().toISOString().slice(0, 7);
-}
-
-function sortGoals(goals: Goal[]) {
-  return [...goals].sort((a, b) => a.priority - b.priority);
-}
-
-const goalTypeLabels: Record<GoalType, string> = {
-  emergency_fund: "Emergency Fund",
-  rent_buffer: "Rent Buffer",
-  school_fees: "School Fees",
-  land_savings: "Land Savings",
-  business_capital: "Business Capital",
-  education: "Education",
-  house_construction: "House / Construction",
-};
 
 export function GoalsWorkspace() {
   const [profile, setProfile] = useState<UserProfile | null>(null);
@@ -122,9 +72,7 @@ export function GoalsWorkspace() {
         linkedAccountId: c.linkedAccountId || reconciledAccounts[0]?.id || "",
       }));
     } catch (loadError) {
-      setError(
-        loadError instanceof Error ? loadError.message : "Unable to load goals.",
-      );
+      setError(loadError instanceof Error ? loadError.message : "Unable to load goals.");
     } finally {
       setIsLoading(false);
     }
@@ -136,7 +84,7 @@ export function GoalsWorkspace() {
     });
   }, []);
 
-  const currentMonth = getCurrentMonth();
+  const currentMonth = new Date().toISOString().slice(0, 7);
   const monthlySummary = useMemo(
     () => getMonthSummary(transactions, [], currentMonth),
     [currentMonth, transactions],
@@ -152,10 +100,10 @@ export function GoalsWorkspace() {
     setError(null);
 
     try {
-      const timestamp = buildTimestamp();
+      const timestamp = new Date().toISOString();
       const goalId = editingGoalId ?? `goal:${crypto.randomUUID()}`;
 
-      const nextGoal: Goal = {
+      await repositories.goals.upsert({
         id: goalId,
         userId: profile.id,
         name: goalForm.name.trim(),
@@ -167,16 +115,13 @@ export function GoalsWorkspace() {
         linkedAccountId: goalForm.linkedAccountId || undefined,
         createdAt: goals.find((g) => g.id === goalId)?.createdAt ?? timestamp,
         updatedAt: timestamp,
-      };
+      });
 
-      await repositories.goals.upsert(nextGoal);
       setGoalForm({ ...defaultGoalForm, linkedAccountId: accounts[0]?.id ?? "" });
       setEditingGoalId(null);
       await loadWorkspace();
     } catch (submitError) {
-      setError(
-        submitError instanceof Error ? submitError.message : "Unable to save goal.",
-      );
+      setError(submitError instanceof Error ? submitError.message : "Unable to save goal.");
     } finally {
       setIsSubmitting(false);
     }
@@ -197,6 +142,7 @@ export function GoalsWorkspace() {
   async function handleDeleteGoal(goalId: string) {
     setIsSubmitting(true);
     setError(null);
+
     try {
       await repositories.goals.remove(goalId);
       if (editingGoalId === goalId) {
@@ -205,9 +151,7 @@ export function GoalsWorkspace() {
       }
       await loadWorkspace();
     } catch (deleteError) {
-      setError(
-        deleteError instanceof Error ? deleteError.message : "Unable to delete goal.",
-      );
+      setError(deleteError instanceof Error ? deleteError.message : "Unable to delete goal.");
     } finally {
       setIsSubmitting(false);
     }
@@ -261,264 +205,28 @@ export function GoalsWorkspace() {
 
       {!isLoading && profile ? (
         <div className="grid gap-5 lg:grid-cols-[0.9fr_1.1fr]">
-          <Card className="border-border/40 shadow-none">
-            <CardHeader>
-              <CardTitle className="text-base">
-                {editingGoalId ? "Edit goal" : "New goal"}
-              </CardTitle>
-              <CardDescription>
-                {goalForm.goalType === "emergency_fund"
-                  ? `Suggested target based on 3x monthly outflow: ${formatCurrency(emergencyFundSuggestion)}`
-                  : "Set a target amount and deadline for this goal."}
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <form className="grid gap-4" onSubmit={handleGoalSubmit}>
-                <div className="grid gap-2">
-                  <Label htmlFor="goal-name">Goal name</Label>
-                  <Input
-                    id="goal-name"
-                    value={goalForm.name}
-                    onChange={(e) => setGoalForm((c) => ({ ...c, name: e.target.value }))}
-                    placeholder="e.g. Emergency savings"
-                    required
-                  />
-                </div>
+          <GoalForm
+            accounts={accounts}
+            goalTypes={defaultGoalTypes}
+            form={goalForm}
+            editingId={editingGoalId}
+            isSubmitting={isSubmitting}
+            emergencyFundSuggestion={emergencyFundSuggestion}
+            onFormChange={setGoalForm}
+            onSubmit={(e) => void handleGoalSubmit(e)}
+            onCancelEdit={() => {
+              setEditingGoalId(null);
+              setGoalForm({ ...defaultGoalForm, linkedAccountId: accounts[0]?.id ?? "" });
+            }}
+          />
 
-                <div className="grid gap-2">
-                  <Label htmlFor="goal-type">Goal type</Label>
-                  <Select
-                    value={goalForm.goalType}
-                    onValueChange={(value) =>
-                      setGoalForm((c) => ({ ...c, goalType: value as GoalType }))
-                    }
-                  >
-                    <SelectTrigger id="goal-type">
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {defaultGoalTypes.map((type) => (
-                        <SelectItem key={type} value={type}>
-                          {goalTypeLabels[type]}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </div>
-
-                <div className="grid gap-2">
-                  <Label htmlFor="target-amount">Target amount (UGX)</Label>
-                  <div className="flex gap-2">
-                    <Input
-                      id="target-amount"
-                      inputMode="decimal"
-                      value={goalForm.targetAmount}
-                      onChange={(e) =>
-                        setGoalForm((c) => ({ ...c, targetAmount: e.target.value }))
-                      }
-                      required
-                    />
-                    {goalForm.goalType === "emergency_fund" && emergencyFundSuggestion > 0 ? (
-                      <Button
-                        type="button"
-                        variant="outline"
-                        size="sm"
-                        className="shrink-0 text-xs"
-                        onClick={() =>
-                          setGoalForm((c) => ({
-                            ...c,
-                            targetAmount: String(Math.round(emergencyFundSuggestion)),
-                          }))
-                        }
-                      >
-                        Use suggested
-                      </Button>
-                    ) : null}
-                  </div>
-                </div>
-
-                <div className="grid gap-2">
-                  <Label htmlFor="target-date">Target date</Label>
-                  <Input
-                    id="target-date"
-                    type="date"
-                    value={goalForm.targetDate}
-                    onChange={(e) =>
-                      setGoalForm((c) => ({ ...c, targetDate: e.target.value }))
-                    }
-                    required
-                  />
-                </div>
-
-                <div className="grid gap-2">
-                  <Label htmlFor="goal-priority">Priority (1 = highest)</Label>
-                  <Input
-                    id="goal-priority"
-                    inputMode="numeric"
-                    min="1"
-                    max="10"
-                    value={goalForm.priority}
-                    onChange={(e) =>
-                      setGoalForm((c) => ({ ...c, priority: e.target.value }))
-                    }
-                    required
-                  />
-                </div>
-
-                {accounts.length > 0 ? (
-                  <div className="grid gap-2">
-                    <Label htmlFor="linked-account">Linked savings account</Label>
-                    <Select
-                      value={goalForm.linkedAccountId || "__none__"}
-                      onValueChange={(value) =>
-                        setGoalForm((c) => ({
-                          ...c,
-                          linkedAccountId: value === "__none__" ? "" : value,
-                        }))
-                      }
-                    >
-                      <SelectTrigger id="linked-account">
-                        <SelectValue placeholder="Any account" />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="__none__">Any account</SelectItem>
-                        {accounts.map((account) => (
-                          <SelectItem key={account.id} value={account.id}>
-                            {account.name}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                  </div>
-                ) : null}
-
-                <div className="flex flex-wrap gap-2">
-                  <Button disabled={isSubmitting} type="submit" size="sm">
-                    {isSubmitting
-                      ? "Saving..."
-                      : editingGoalId
-                        ? "Update goal"
-                        : "Create goal"}
-                  </Button>
-                  {editingGoalId ? (
-                    <Button
-                      type="button"
-                      variant="outline"
-                      size="sm"
-                      onClick={() => {
-                        setEditingGoalId(null);
-                        setGoalForm({ ...defaultGoalForm, linkedAccountId: accounts[0]?.id ?? "" });
-                      }}
-                    >
-                      Cancel
-                    </Button>
-                  ) : null}
-                </div>
-              </form>
-            </CardContent>
-          </Card>
-
-          <Card className="border-border/40 shadow-none">
-            <CardHeader>
-              <CardTitle className="text-base">Your goals</CardTitle>
-              <CardDescription>
-                Monthly contribution targets are calculated from target, deadline, and current progress.
-              </CardDescription>
-            </CardHeader>
-            <CardContent className="grid gap-3">
-              {goals.length === 0 ? (
-                <div className="rounded-md border border-dashed border-border/50 px-4 py-8 text-sm text-muted-foreground">
-                  No goals yet. Create your first goal to start building.
-                </div>
-              ) : (
-                goals.map((goal) => {
-                  const plan = getGoalContributionPlan(goal);
-                  const linkedAccount = accounts.find((a) => a.id === goal.linkedAccountId);
-                  const progressPercent =
-                    goal.targetAmount > 0
-                      ? Math.min(100, Math.round((goal.currentAmount / goal.targetAmount) * 100))
-                      : 0;
-
-                  return (
-                    <Card
-                      key={goal.id}
-                      className="border-border/40 bg-muted/30 shadow-none"
-                    >
-                      <CardContent className="px-4 py-4">
-                        <div className="flex items-start justify-between gap-3">
-                          <div className="space-y-0.5">
-                            <div className="text-sm font-medium text-foreground">
-                              {goal.name}
-                            </div>
-                            <div className="text-xs text-muted-foreground">
-                              {goalTypeLabels[goal.goalType]}
-                              {linkedAccount ? ` · ${linkedAccount.name}` : ""}
-                            </div>
-                          </div>
-                          <div className="flex gap-1.5">
-                            <Button
-                              size="sm"
-                              variant="ghost"
-                              className="h-7 text-xs"
-                              onClick={() => beginGoalEdit(goal)}
-                            >
-                              Edit
-                            </Button>
-                            <Button
-                              size="sm"
-                              variant="ghost"
-                              className="h-7 text-xs text-destructive hover:text-destructive"
-                              onClick={() => void handleDeleteGoal(goal.id)}
-                            >
-                              Delete
-                            </Button>
-                          </div>
-                        </div>
-
-                        <div className="mt-3">
-                          <div className="mb-1.5 flex items-center justify-between text-xs text-muted-foreground">
-                            <span>{formatCurrency(goal.currentAmount)}</span>
-                            <span>{progressPercent}% of {formatCurrency(goal.targetAmount)}</span>
-                          </div>
-                          <div className="h-1.5 w-full overflow-hidden rounded-full bg-muted">
-                            <div
-                              className="h-full rounded-full bg-primary transition-all"
-                              style={{ width: `${progressPercent}%` }}
-                            />
-                          </div>
-                        </div>
-
-                        <Separator className="my-3" />
-
-                        <div className="grid grid-cols-2 gap-3 text-sm">
-                          <div>
-                            <div className="text-xs text-muted-foreground">Monthly target</div>
-                            <div className="font-medium tabular-nums">
-                              {formatCurrency(plan.monthlyContribution)}
-                            </div>
-                          </div>
-                          <div>
-                            <div className="text-xs text-muted-foreground">Deadline</div>
-                            <div className="font-medium">{goal.targetDate}</div>
-                          </div>
-                        </div>
-
-                        {plan.isBehindSchedule ? (
-                          <p className="mt-3 text-xs text-amber-600 dark:text-amber-400">
-                            This goal is behind schedule — increase contributions to stay on track.
-                          </p>
-                        ) : (
-                          <p className="mt-3 text-xs text-muted-foreground">
-                            {plan.monthsRemaining} month{plan.monthsRemaining !== 1 ? "s" : ""} remaining.
-                          </p>
-                        )}
-                      </CardContent>
-                    </Card>
-                  );
-                })
-              )}
-            </CardContent>
-          </Card>
+          <GoalList
+            accounts={accounts}
+            goals={goals}
+            isSubmitting={isSubmitting}
+            onEdit={beginGoalEdit}
+            onDelete={(id) => void handleDeleteGoal(id)}
+          />
         </div>
       ) : null}
     </div>
